requirements.txt
flask~=0.12
flask-admin~=1.5
flask-cors~=3.0
flask-login~=0.4
flask-marshmallow~=0.8
flask-migrate~=2.1
flask-restful~=0.3
flask-sqlalchemy~=2.3
geoalchemy2~=0.4
marshmallow-sqlalchemy~=0.13
psycopg2~=2.7
rauth~=0.7
shapely~=1.6